+++ CHANGELOG.rst
@@ -6,6 +6,31 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
+0.5.0
+--------
+
+Added
+++++++
+- The ability to fit **multiple** latent phenotypes in the global epistasis models. This adds the ``n_latent_phenotypes`` flag to ``AbstractEpistasis`` models, and changes calls to certain methods / properties of that abstract model class and its concrete subclasses.
+
+- The concept of "bottleneck" likelihoods in global epistasis models, implemented in ``BottleneckLikelihood``.
+
+- The ``bottlenecks`` module to estimate bottlenecks.
+
+- Added ``AbstractEpistasis.aic`` property.
+
+- Added ``globalepistasis.fit_models``
+
+- Added ``MultiLatentSigmoidPhenotypeSimulator``.
+
+- An equals (``__eq__``) comparison operation to ``BinaryMap``.
+
+- Added ``n_pre`` and ``n_post`` attributes to ``BinaryMap``. This changes the initialization to add new parameters, ``n_pre_col``, ``n_post_col``, and ``cols_optional``.
+
+Fixed
+++++++
+- ``BinaryMap`` objects can now be deep copied (they don't have a compiled regex as attributed).
+
 0.4.7
 ------